# gin-swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "这里填写文档名称(必填项): Gin-Swagger在线文档测试项目",
        "termsOfService": "这里填写服务条款: http://swagger.io/terms/",
        "contact": {},
        "version": "0.0.1(必填项)"
    },
    "host": "localhost:8080(服务器运行地址)",
    "basePath": "/api(API基本路径)",
    "paths": {
        "/api/depart": {
            "post": {
                "description": "添加部门信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "添加部门",
                "parameters": [
                    {
                        "description": "部门名称等信息",
                        "name": "addBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.AddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/depart/{name}": {
            "get": {
                "description": "查询部门信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "查询部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "部门名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "description": "删除部门信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "部门名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "patch": {
                "description": "修改部门信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "修改部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "部门名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户列表",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "description": "用户通过用户名/密码登录系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "登录API",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "loginBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rsp.Login"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "注册API",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "registerBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rsp.Register"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/user/{username}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询用户信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "查询API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rsp.Information"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除用户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "删除API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "修改API",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "modifyBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.ModifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
        "form.AddRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "部门名称",
                    "type": "string"
                }
            }
        },
        "form.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "form.ModifyRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "city": {
                    "description": "城市",
                    "type": "string"
                },
                "sex": {
                    "description": "性别",
                    "type": "integer"
                }
            }
        },
        "form.RegisterRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "city": {
                    "description": "城市",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "sex": {
                    "description": "性别",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "rsp.Information": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "city": {
                    "description": "城市",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "sex": {
                    "description": "性别",
                    "type": "integer"
                }
            }
        },
        "rsp.Login": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Token令牌",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "rsp.Register": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "city": {
                    "description": "城市",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "sex": {
                    "description": "性别",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API安全验证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}